EY-2646: Vi har så mange grunnlag-mapping-metodar her at dei fortener si eiga fil
