+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/behandling/Grunnlagmapper.kt
@@ -1,0 +1,83 @@
+package no.nav.etterlatte.brev.behandling
+
+import no.nav.etterlatte.brev.model.Spraak
+import no.nav.etterlatte.libs.common.behandling.Persongalleri
+import no.nav.etterlatte.libs.common.behandling.SakType
+import no.nav.etterlatte.libs.common.grunnlag.Grunnlag
+import no.nav.etterlatte.libs.common.grunnlag.hentDoedsdato
+import no.nav.etterlatte.libs.common.grunnlag.hentFoedselsnummer
+import no.nav.etterlatte.libs.common.grunnlag.hentKonstantOpplysning
+import no.nav.etterlatte.libs.common.grunnlag.hentNavn
+import no.nav.etterlatte.libs.common.grunnlag.hentVergemaalellerfremtidsfullmakt
+import no.nav.etterlatte.libs.common.grunnlag.opplysningstyper.Navn
+import no.nav.etterlatte.libs.common.grunnlag.opplysningstyper.Opplysningstype
+import no.nav.pensjon.brevbaker.api.model.Foedselsnummer
+
+fun Grunnlag.mapSoeker(): Soeker =
+    with(this.soeker) {
+        val navn = hentNavn()!!.verdi
+
+        Soeker(
+            fornavn = navn.fornavn.storForbokstav(),
+            mellomnavn = navn.mellomnavn?.storForbokstav(),
+            etternavn = navn.etternavn.storForbokstav(),
+            fnr = Foedselsnummer(hentFoedselsnummer()!!.verdi.value),
+        )
+    }
+
+fun Grunnlag.mapAvdoed(): Avdoed =
+    with(this.familie) {
+        val avdoed = hentAvdoed()
+
+        Avdoed(
+            navn = avdoed.hentNavn()!!.verdi.fulltNavn(),
+            doedsdato = avdoed.hentDoedsdato()!!.verdi!!,
+        )
+    }
+
+fun Grunnlag.mapInnsender(): Innsender? =
+    with(this.sak) {
+        val opplysning = hentKonstantOpplysning<Persongalleri>(Opplysningstype.PERSONGALLERI_V1)
+
+        val persongalleri =
+            requireNotNull(opplysning?.verdi) {
+                "Sak (id=${metadata.sakId}) mangler opplysningstype PERSONGALLERI_V1"
+            }
+
+        persongalleri.innsender?.let {
+            Innsender(fnr = Foedselsnummer(it))
+        }
+    }
+
+fun Grunnlag.mapSpraak(): Spraak =
+    with(this.sak) {
+        val opplysning = hentKonstantOpplysning<Spraak>(Opplysningstype.SPRAAK)
+
+        requireNotNull(opplysning?.verdi) {
+            "Sak (id=${metadata.sakId}) mangler opplysningstype SPRAAK"
+        }
+    }
+
+fun Grunnlag.mapVerge(sakType: SakType): Verge? =
+    with(this) {
+        val opplysning = sak.hentVergemaalellerfremtidsfullmakt()
+
+        if (opplysning?.verdi != null) {
+            TODO("Støtter ikke annen verge enn forelder – håndtering av annen verge krever ytterligere avklaringer")
+        } else if (sakType == SakType.BARNEPENSJON) {
+            val gjenlevendeNavn = hentGjenlevende().hentNavn()!!.verdi.fulltNavn()
+
+            Verge(gjenlevendeNavn)
+        } else {
+            null
+        }
+    }
+
+private fun Navn.fulltNavn(): String = listOfNotNull(fornavn, mellomnavn, etternavn).joinToString(" ") { it.storForbokstav() }
+
+private fun String.storForbokstav() = this.lowercase().storForbokstavEtter("-").storForbokstavEtter(" ")
+
+private fun String.storForbokstavEtter(delim: String) =
+    this.split(delim).joinToString(delim) {
+        it.replaceFirstChar { c -> c.uppercase() }
+    }
